GG Settings Role Attached
This picture displays the attached Greengrass Roles
To attach the Greengrass Role created at the AIM Roles, go to the end of the following location and select "Attach role" to attach the roles.
IoT > Settings > Greengrass Roles

AWS Info note: Greengrass needs your permission to access these services and read and write data on your behalf. The default permissions are described in the AWSGreengrassResourceAccessRolePolicy.